pyproject.toml
[project]
name = "llmstart-homework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.21.0",
    "openai>=1.99.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "structlog>=25.4.0",
]
